requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
xgboost
mlflow
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
pyyaml
sentence-transformers
faiss-cpu
matplotlib
seaborn
pytest
httpx
